[Refactor:InstallScript] Update command-line arguments and add skip options for node and npm build

+++ .setup/install_submitty/install_site.sh
@@ -11,9 +11,12 @@
 #   - if composer.json or composer.lock was modified, run composer
 #   - only set permissions for modified files
 #
-# This script has two command-line arguments:
+# This script has five command-line arguments:
 #   Required argument: config=<config dir>.
 #   Optional argument: browscap
+#   Optional argument: skip-vendor
+#   Optional argument: skip-node
+#   Optional argument: skip-npm-build
 
 set_permissions () {
     local fullpath=$1
@@ -88,7 +91,7 @@ done
 
 if [ -z "${SUBMITTY_CONFIG_DIR}" ]; then
     echo "ERROR: This script requires a config dir argument"
-    echo "Usage: ${BASH_SOURCE[0]} config=<config dir> [browscap]"
+    echo "Usage: ${BASH_SOURCE[0]} config=<config dir> [browscap] [skip-vendor] [skip-node] [skip-npm-build]"
     exit 1
 fi
 
@@ -121,15 +124,18 @@ chmod 755 ${SUBMITTY_DATA_DIR}/run
 chown ${PHP_USER}:www-data ${SUBMITTY_DATA_DIR}/run/websocket
 chmod 2750 ${SUBMITTY_DATA_DIR}/run/websocket
 
-# Delete all typescript code to prevent deleted files being left behind and potentially
-# causing compilation errors
-if [ -d "${SUBMITTY_INSTALL_DIR}/site/ts" ]; then
-    rm -r "${SUBMITTY_INSTALL_DIR}/site/ts"
-fi
 
-# Delete all vue code to prevent deleted vue files from being rendered
-if [ -d "${SUBMITTY_INSTALL_DIR}/site/vue" ]; then
-    rm -r "${SUBMITTY_INSTALL_DIR}/site/vue"
+if [[ -z "${SKIP_NODE}" ]]; then
+    # Delete all typescript code to prevent deleted files being left behind and potentially
+    # causing compilation errors
+    if [ -d "${SUBMITTY_INSTALL_DIR}/site/ts" ]; then
+        rm -r "${SUBMITTY_INSTALL_DIR}/site/ts"
+    fi
+
+    # Delete all vue code to prevent deleted vue files from being rendered
+    if [ -d "${SUBMITTY_INSTALL_DIR}/site/vue" ]; then
+        rm -r "${SUBMITTY_INSTALL_DIR}/site/vue"
+    fi
 fi
 
 # copy the website from the repo. We don't need the tests directory in production and then
@@ -202,12 +208,6 @@ fi
 # create lang cache directory
 mkdir -p ${SUBMITTY_INSTALL_DIR}/site/cache/lang
 
-if [ -d "${SUBMITTY_INSTALL_DIR}/site/public/mjs" ]; then
-    rm -r "${SUBMITTY_INSTALL_DIR}/site/public/mjs"
-fi
-# create output dir for esbuild
-mkdir -p ${SUBMITTY_INSTALL_DIR}/site/public/mjs
-
 # Update ownership to PHP_USER for affected files and folders
 chown ${PHP_USER}:${PHP_GROUP} ${SUBMITTY_INSTALL_DIR}/site
 for entry in "${result_array[@]}"; do
@@ -216,8 +216,6 @@ done
 
 # Update ownership for cgi-bin to CGI_USER
 find ${SUBMITTY_INSTALL_DIR}/site/cgi-bin -exec chown ${CGI_USER}:${CGI_GROUP} {} \;
-
-chown ${PHP_USER}:${PHP_GROUP} ${SUBMITTY_INSTALL_DIR}/site/public/mjs
 
 # set the mask for composer so that it'll run properly and be able to delete/modify
 # files under it
@@ -413,6 +411,14 @@ mkdir -p "${SUBMITTY_INSTALL_DIR}/site/incremental_build"
 chgrp "${PHP_USER}" "${SUBMITTY_INSTALL_DIR}/site/incremental_build"
 
 if [[ -z "${SKIP_NPM_BUILD}" ]]; then
+    if [ -d "${SUBMITTY_INSTALL_DIR}/site/public/mjs" ]; then
+        rm -r "${SUBMITTY_INSTALL_DIR}/site/public/mjs"
+    fi
+    # create output dir for esbuild
+    mkdir -p ${SUBMITTY_INSTALL_DIR}/site/public/mjs
+    chown ${PHP_USER}:${PHP_GROUP} ${SUBMITTY_INSTALL_DIR}/site/public/mjs
+
+
     echo "Running esbuild"
     chmod a+x ${NODE_FOLDER}/esbuild/bin/esbuild
     chmod a+x ${NODE_FOLDER}/typescript/bin/tsc
